+++ src/Ydb.Sdk/src/Auth/ICredentialsProvider.cs
@@ -1,13 +1,11 @@
-﻿using Ydb.Sdk.Services.Auth;
-
-namespace Ydb.Sdk.Auth;
+﻿namespace Ydb.Sdk.Auth;
 
 public interface ICredentialsProvider
 {
-    // For removal in 1.*
-    string? GetAuthInfo();
+    ValueTask<string> GetAuthInfoAsync();
+}
 
-    ValueTask<string?> GetAuthInfoAsync() => ValueTask.FromResult(GetAuthInfo());
-
-    Task ProvideAuthClient(AuthClient authClient) => Task.CompletedTask;
+public interface IAuthClient
+{
+    Task<TokenResponse> FetchToken();
 }